\name{print.summary.vdracox}
\alias{print.summary.vdracox}
\title{
	Produces a printed summary of a fitted vdra cox model.
}
\description{
}
\usage{
## S3 method for class 'summary.vdracox'
print(x)
}
\arguments{
	\item{x}{the result of a call to summary.vdracox}
}
\seealso{
  \code{\link{summary.vdracox}}
}
