\name{print.summary.vdralinear}
\alias{print.summary.vdralinear}
\title{
	Produces a printed summary of a fitted vdra linear regression model.
}
\description{
}
\usage{
## S3 method for class 'summary.vdralinear'
print(x, lion = FALSE)
}
\arguments{
	\item{x}{the result of a call to summary.vdralinear}
	\item{lion}{logical.  If \code{TRUE}, prints the PSU Nittany lion to the right of the summary statistics}
}
\seealso{
  \code{\link{summary.vdralinear}}
}
