\name{print.summary.vdralogistic}
\alias{print.summary.vdralogistic}
\title{
	Produces a printed summary of a fitted vdra logistic regression model.
}
\description{
}
\usage{
## S3 method for class 'summary.vdralogistic'
print(x, lion = FALSE)
}
\arguments{
	\item{x}{the result of a call to summary.vdralogistic}
	\item{lion}{logical.  If \code{TRUE}, prints the PSU Nittany lion to the right of the summary statistics}
}
\seealso{
  \code{\link{summary.vdralogistic}}
}
