% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertical_utilities.R
\name{summary.vdralogistic}
\alias{summary.vdralogistic}
\alias{summary.vdralogistic.object}
\alias{print.summary.vdralogistic}
\title{Summary Method for Vertical Distributed Logistic Regression Models}
\usage{
\method{summary}{vdralogistic}(object, ...)
}
\arguments{
\item{object}{a \code{vdralogistic} object.}

\item{...}{futher argumetns passed to or from other methods.}
}
\value{
Returns an object of class \code{summary.vdralogistic}. Objects of
  this class have a method for the function \code{print}.  The following
  components must be included in \code{summary.vdralogistic} object.
  \describe{

  \item{failed}{ogical value.  If \code{FALSE}, then there was an error
  processing the data.  if \code{TRUE}, there were no errors.}

  \item{converged}{ogical value.  If \code{TRUE}, the regression converged.
  If \code{FALSE}, it did not.}

  \item{party}{ vector which indicates the party from which each covariate
  came.}

  \item{coefficients}{he vector of coefficients.  If the model is
  over-determined, there will be missing values in the vector corresponding
  to the redudant columns model matrix.}

  \item{secoef}{he vector of the standard error of the coefficients.}

  \item{tvals}{he t-values of the coefficietns.}

  \item{pvals}{he p-values of the coefficients.}

  \item{nulldev}{he null deviance of the fit.}

  \item{nulldev_df}{he degrees of freedom for the null deviance.}

  \item{resdev}{he residual deviance of the fit.}

  \item{resdev_df}{he degrees of freedome for the residual deviance.}

  \item{aic the}{IC of the fit.}

  \item{bic the}{IC of the fit.}

  \item{iter }{ number of iterations of the cox algorithm before
  convergence.}

  }
}
\description{
Produces a summary of a fitted vdra logistic regression model.
}
\examples{
summary(vdra_fit_logistic_A)
}
\seealso{
\code{\link{vdralogistic}}
}
