% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdra_data.R
\docType{package}
\name{vdra}
\alias{vdra}
\title{vdra: A package for performing vertical distributed regression analysis.}
\description{
All the functions in this package are desingned to be run in a distributed
network using popMedNet.
}
