% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertical_utilities.R
\name{RocTest}
\alias{RocTest}
\alias{rocdistributed.object}
\alias{print.rocdistributed}
\title{Create the ROC for Vertical Distributed Logistic Regression}
\usage{
RocTest(x = NULL, bins = 10)
}
\arguments{
\item{x}{an object of type \code{\link{vdralogistic}}.}

\item{bins}{the number of bins the data will be separated into.}
}
\value{
Returns an object of class \code{rocdistributed}. Objects of this
  class have a method for the function \code{print}. The following components
  must be included in a \code{rocdistributed} object.
\describe{
 \item{roc}{a two column matrix containing the coordinates of 1 - specificity
 and sensitivity.}
 \item{auc}{numeric value which is area under the curve.}
}
}
\description{
Generate the receiver operator curve on an object created by
  2-party, 3-party, or K-party vdra logistic regression.  Only the party that
  holds the response may invoke this function.
}
\examples{
RocTest(vdra_fit_logistic_A)

RocTest(vdra_fit_logistic_A, 40)
}
